Refactor Awesome Books with ES6 Syntax
