+++ index.html
@@ -19,7 +19,7 @@
           <li><a id="link-contact" href="#">Contact</a></li>
         </ul>
       </nav>
-      <p class="date-section">January 12th 2023, 07:15:44 am</p>
+      <p class="date-section"></p>
     </header>
   
     <main>
@@ -42,7 +42,7 @@
         <h2 class="contact-headline">Contact Information</h2>
         <p class="contact-info">Do have any question or you just want to say "Hello"?<br>You can reach out to us!</p>
         <ul class="contact-list">
-          <li>Our Email: dhirajsingha@mail.com</li>
+          <li>Our Email: shaaibusuleiman9@gmail.com</li>
           <li>Our phone no: +2347039606814</li>
           <li>Our address: Hollowway road 22,84503, London, UK</li>
         </ul>
@@ -52,6 +52,6 @@
     <footer class="footer">
       <p>&copy;copyright 2023</p>
     </footer>
-    <script src="index.js"></script>
+    <script type="module" src="index.js"></script>
 </body>
 </html>
